+++ ksa_compliance/hooks.py
@@ -134,7 +134,8 @@ jinja = {
 doc_events = {
     "Sales Invoice": {
         "on_submit": "ksa_compliance.standard_doctypes.sales_invoice.create_sales_invoice_additional_fields_doctype",
-        "validate": "ksa_compliance.standard_doctypes.sales_invoice.validate_tax_category",
+        "validate": ["ksa_compliance.standard_doctypes.sales_invoice.validate_tax_category",
+                     "ksa_compliance.standard_doctypes.sales_invoice.validate_tax_rate",],
         "before_cancel": "ksa_compliance.standard_doctypes.sales_invoice.prevent_cancellation_of_sales_invoice"
     },
 }